Extracted icon colorization logic
